Update is working fine. Patient Cases and Appointment only delete when there is no child.

+++ client/src/app/components/view-all-appointments/view-all-appointments.component.ts
@@ -22,9 +22,12 @@ export class ViewAllAppointmentsComponent {
     )
   }
   deleteAppointment(id:any){
-    this.appointService.deleteAppointment(id)
-    .subscribe(
-      ()=>this.ngOnInit()
-    )
-  }
+    let isExecuted = confirm("Are you sure you want to delete this appointment?")
+    if(isExecuted){
+      this.appointService.deleteAppointment(id)
+      .subscribe(
+        ()=>this.ngOnInit()
+        )
+      }
+    }
 }
